requirements.txt
aiohttp
aiomysql
python-dotenv
mysql-connector-python
aiofiles
openai
fastapi
uvicorn
pydantic